data_pipelines/pyproject.toml
[project]
name = "data_pipelines"
version = "0.1.0"
dependencies = [
    "apache-airflow==3.1.8",
    "duckdb>=1.5.0",
]
requires-python = ">=3.12,<3.14"
authors = [
    {name = "Daniel Schruhl", email = "danielschruhl@gmail.com"},
]
description = "Data pipeline and orchestration"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.6",
]

[tool.uv]
constraint-dependencies = [
    "apache-airflow==3.1.8",
]

[project.scripts]
data_pipeline = "data_pipeline.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
